requirements.txt
pdfplumber==0.5.24

autopep8
flake8

requests==2.25.0dw
requests-ouathlib==1.3.0
python-dotenv==0.15.0